03_deepeval_framework/requirements.txt
deepeval>=2.6.0
openai>=1.54
groq>=0.11
instructor>=1.6
pydantic>=2.9
pytest>=8.3
pytest-html>=4.1
pytest-xdist>=3.6
requests>=2.32
python-dotenv>=1.0
grpcio>=1.66
